+++ .github/workflows/validate.yml
@@ -30,7 +30,8 @@ jobs:
           job-name-prefix: 'Build' # Optional, defaults to empty string
       - name: Verify Job Output was generated
         run: |
-          if [ -z "${{ steps.setup-jobs.outputs.jobs }}" ]; then
+          output=$(echo "${{ steps.setup-jobs.outputs.jobs }}" | jq -r '. | length')
+          if [ -z "$output" ] || [ "$output" -eq 0 ]; then
             echo "No jobs were generated. Please check your build options.";
             exit 1;
           fi
